dist_pyc/requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
supabase>=2.0.0
qrcode>=7.4.2
Pillow>=10.0.0
uvloop>=0.19.0;platform_system!="Windows"
